requirements.txt
streamlit
supabase
google-generativeai
pandas
python-dotenv
